requirements.txt
numpy==1.26.4
scipy==1.12.0
web3==6.15.1
python-dotenv==1.0.1
